# openapi/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Ads Marketplace API",
    "version": "1.0.0",
    "description": "API documentation for the Ads Marketplace platform - connecting advertisers with channel owners on Telegram",
    "contact": {
      "name": "API Support"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Development server"
    }
  ],
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT authorization token"
      },
      "TelegramAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Telegram-Init-Data",
        "description": "Telegram Mini App init data"
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "description": "Error message"
          },
          "details": {
            "type": "object",
            "description": "Additional error details"
          }
        }
      },
...
    }
  },
...
  "paths": {
...
    "/api/auth/telegram": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Authenticate via Telegram Mini App",
        "description": "Authenticates a user using Telegram Mini App init data and returns a JWT token",
        "security": [
          {
            "TelegramAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully authenticated",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "token": {
                      "type": "string",
                      "description": "JWT authentication token"
                    },
                    "user": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "string"
                        },
                        "telegramId": {
                          "type": "string"
                        },
                        "username": {
                          "type": "string",
                          "nullable": true
                        },
                        "firstName": {
                          "type": "string",
                          "nullable": true
                        },
                        "lastName": {
                          "type": "string",
                          "nullable": true
                        },
                        "isAdvertiser": {
                          "type": "boolean"
                        },
                        "isChannelOwner": {
                          "type": "boolean"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Invalid or missing Telegram init data",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/wallet/connect": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Connect TON wallet to user account",
        "description": "Links a TON wallet address to the authenticated user's account",
        "security": [
          {
            "TelegramAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "address"
                ],
                "properties": {
                  "address": {
                    "type": "string",
                    "description": "TON wallet address",
                    "example": "EQD..."
                  },
                  "publicKey": {
                    "type": "string",
                    "description": "Wallet public key"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Wallet successfully connected",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "wallet": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "string"
                        },
                        "address": {
                          "type": "string"
                        },
                        "isActive": {
                          "type": "boolean"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid request or wallet already connected",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/wallet/verify": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Verify wallet ownership via signature",
        "description": "Verifies wallet ownership by validating a signature of a message",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "address",
                  "signature",
                  "message"
                ],
                "properties": {
                  "address": {
                    "type": "string",
                    "description": "TON wallet address to verify"
                  },
                  "signature": {
                    "type": "string",
                    "description": "Signed message signature"
                  },
                  "message": {
                    "type": "string",
                    "description": "Original message that was signed"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Wallet verified successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "verified": {
                      "type": "boolean"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid signature or wallet not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/wallet/{address}": {
      "delete": {
        "tags": [
          "Auth"
        ],
        "summary": "Disconnect a wallet from user account",
        "description": "Removes the wallet connection from the authenticated user's account",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "address",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "TON wallet address to disconnect"
          }
        ],
        "responses": {
          "200": {
            "description": "Wallet disconnected successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Wallet not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
  }
}